Moving comments into separate component.
