+++ timesketch/frontend-ng/src/components/Explore/EventList.vue
@@ -366,9 +366,38 @@ limitations under the License.
 
         <!-- Comment field -->
         <template v-slot:item._source.comment="{ item }">
-          <v-badge :offset-y="10" bordered v-if="item._source.comment.length" :content="item._source.comment.length">
-            <v-icon small @click="toggleDetailedEvent(item)"> mdi-comment-text-multiple-outline </v-icon>
-          </v-badge>
+          <v-tooltip top open-delay="300">
+            <template v-slot:activator="{ on }">
+              <div v-on="on" class="d-inline-block">
+                <v-btn
+                  icon
+                  small
+                  @click="toggleDetailedEvent(item)"
+                   v-if="item._source.comment.length"
+                >
+                  <v-badge :offset-y="10" :offset-x="10" bordered :content="item._source.comment.length">
+                    <v-icon small> mdi-comment-text-multiple-outline </v-icon>
+                  </v-badge>
+                </v-btn>
+              </div>
+            </template>
+            <span>Open comments</span>
+          </v-tooltip>
+          <v-tooltip top open-delay="300">
+            <template v-slot:activator="{ on }">
+              <div v-on="on" class="d-inline-block">
+                <v-btn
+                  icon
+                  small
+                  @click="toggleDetailedEvent(item)"
+                  v-if="item['showDetails'] && !item._source.comment.length"
+                >
+                  <v-icon> mdi-comment-plus-outline </v-icon>
+                </v-btn>
+              </div>
+            </template>
+            <span>Add a comment</span>
+          </v-tooltip>
         </template>
       </v-data-table>
     </div>
